fix(brute-worker): cast i64 to f64 before D1 bind to avoid BigInt error
wasm-bindgen converts Rust i64 to JS BigInt, which Cloudflare D1 rejects
with D1_TYPE_ERROR. Cast all i64 values to f64 before .into() across all
D1 bind call sites: timestamps, time-series buckets, limit/since params,
checked_at, cutoff_ms, and domain_total.

+++ brute-worker/src/abuse/mod.rs
@@ -101,7 +101,7 @@ impl AbuseProvider for WorkerAbuseIpDb {
                 ip.into(),
                 confidence_score.into(),
                 total_reports.into(),
-                now.into(),
+                (now as f64).into(),
             ])
             .map_err(|e| BruteError::Database(e.to_string()))?
             .run()
